Fix rubocop class limit
